refactor(slice4): geteiltes CLI-Patch-Util (DRY, duplication-Gate), clip-settings+fades nutzen es
